+++ src/index.ts
@@ -11,6 +11,7 @@ import applyStatusReport from "./statusReport";
 import { startStandaloneServer } from "@apollo/server/standalone";
 import { createApolloServer, gqlContext } from "gql/server";
 import { getEsClient } from "external/elasticsearch";
+import indexingRouter from "routes/indexing";
 
 (async () => {
 	/**
@@ -32,42 +33,13 @@ import { getEsClient } from "external/elasticsearch";
   });
   logger.info(`GQL server ready at port: ${GRAPHQL_PORT}`);
 
-  expressApp.post("/index/program/:program_id", async (req, res) => {
-    const programId = req.params.program_id;
-    try {
-      logger.info(
-        `received request to index program ${programId}, validating program id...`
-      );
-      // validate programId:
-      const regex = new RegExp(
-        "^[A-Z0-9][-_A-Z0-9]{2,7}[-](([A-Z][A-Z])|(INTL))$"
-      );
-      const found = programId.match(regex);
+  expressApp.post("/index", indexingRouter);
 
-			if (!found) {
-				return res
-					.status(400)
-					.send(`ProgramId (${programId}) is invalid, please enter a valid programId.`);
-			} else {
-				await queueProgramUpdateEvent({
-					programId: programId,
-					type: KnownEventType.SYNC,
-					rdpcGatewayUrls: [RDPC_URL],
-				});
-				logger.info(`Program ${programId} has been queued for indexing.`);
-				return res.status(200).send(`Program ${programId} has been queued for indexing.`);
-			}
-		} catch (error) {
-			logger.error('Error in processing index program request: ' + error);
-			return res.status(500).send(`Failed to queue program ${programId} for indexing.`);
-		}
-	});
-
-	// Initialize Kafka Consumers and Producers
-	//  - Can be disabled to simplify running in dev, set FEATURE_DEV_DISABLE_KAFKA=true in .env
-	if (featureFlags.kafka) {
-		await kafka.setup();
-	}
+  // Initialize Kafka Consumers and Producers
+  //  - Can be disabled to simplify running in dev, set FLAG_DEV_DISABLE_KAFKA=true in .env
+  if (featureFlags.kafka) {
+    await kafka.setup();
+  }
 
 	logger.info('pipeline is ready!');
 	expressApp.listen(PORT, () => {
